Proof output in logs is unicode, marshall instead
The output is in unicode and so does not give any meaningful
information.

Add a single marshall to convert into readable JSON

Also commented out the pubkey which was being outputted in DER
format. If there is a need to have this shown in the logs, we can
look to decode.