remove optimization to use client-provided leaf directly
When leaves are added to the merkle tree by rekor, we marshal the leaf which will have a
deterministic ordering given the json.Marshal semantics. If we use the client data
directly, then it may have a different ordering (but still be valid JSON), but the hasher
would generate a different hash and we may not find the leaf in the tree.

This change re-marshals the input data against the RekorLeaf struct to ensure we're using
the same ordering as when the leaf was inserted.