improve memory utilization of API endpoints, error checking (#47)
* improve memory utilization of API endpoints, error checking

this change includes several optimizations to minimize the memory footprint
of the rekor-server, including:

- validating input signature, public key, SHA values are valid before CPU &
memory intensive operations occur before inserting leaf into tLOG
- minimizing or removing copies of data in memory where possible
- parallelize hash computation & signature validation for remote fetch
- detect content-type for remote resource to determine if gzip is needed,
instead of relying on file extension

* remove optimization to use client-provided leaf directly

When leaves are added to the merkle tree by rekor, we marshal the leaf which will have a
deterministic ordering given the json.Marshal semantics. If we use the client data
directly, then it may have a different ordering (but still be valid JSON), but the hasher
would generate a different hash and we may not find the leaf in the tree.

This change re-marshals the input data against the RekorLeaf struct to ensure we're using
the same ordering as when the leaf was inserted.